:art: Create QuestionGeneratorBase, and move question generation logic to it. Refactor FlashCards class to inherit from QuestionGeneratorBase and
